+++ src/components/NavigationSection.svelte
@@ -15,7 +15,7 @@
   {{else}}
     {{#if categories}}
       {{#each categories as category}}
-        <h2 class='coz-nav-category'>{{t(`Categories.${category.title}`)}}</h2>
+        <h2 class='coz-nav-category'>{{t(`Categories.${category.slug}`)}}</h2>
         <NavigationGroup group='{{category.items}}' itemsLimit={{4}} separator='bottom' />
       {{/each}}
     {{else}}
@@ -28,6 +28,7 @@
 
 <script>
   import { t } from '../lib/i18n'
+  import { getCategorizedItems } from '../lib/helpers'
 
   import NavigationGroup from './NavigationGroup'
 
@@ -74,27 +75,6 @@
     } else {
       close.call(this)
     }
-  }
-
-  // Take an items array and return an array of category objects with the matching title and items
-  function getCategorizedItems (items) {
-    if (items[0] instanceof Array) return null // doesn't handle this case
-    const categorizedItemsObject = items.reduce((accumulator, item) => {
-      accumulator[item.category] = accumulator[item.category] || []
-      accumulator[item.category].push(item)
-      return accumulator
-     }, {})
-
-    return Object.keys(categorizedItemsObject)
-      .map(category => {
-        return {title: category, items: categorizedItemsObject[category]}
-      })
-      // categories alphabetical sorting
-      .sort((c1, c2) => {
-        if (t(`Categories.${c1.title}`) > t(`Categories.${c2.title}`)) return 1
-        if (t(`Categories.${c1.title}`) < t(`Categories.${c2.title}`)) return -1
-        return 0
-      })
   }
 
   export default {